+++ s4e-backend/src/main/java/pl/cyfronet/s4e/bean/Scene.java
@@ -27,6 +27,9 @@ public class Scene extends CreationAndModificationAudited {
     public static final String COLUMN_ID = "id";
     public static final String COLUMN_PRODUCT_ID = "product_id";
     public static final String COLUMN_TIMESTAMP = "timestamp";
+    public static final String COLUMN_FOOTPRINT = "footprint";
+    public static final String COLUMN_METADATA = "metadata_content";
+    public static final String COLUMN_CONTENT = "scene_content";
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;