+++ s4e-backend/src/main/java/pl/cyfronet/s4e/service/SearchService.java
@@ -2,9 +2,9 @@ package pl.cyfronet.s4e.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import pl.cyfronet.s4e.bean.Scene;
-import pl.cyfronet.s4e.data.repository.SceneRepository;
+import pl.cyfronet.s4e.api.MappedScene;
 import pl.cyfronet.s4e.api.SearchConverter;
+import pl.cyfronet.s4e.data.repository.SceneRepository;
 
 import java.sql.SQLException;
 import java.util.HashMap;
@@ -17,7 +17,7 @@ public class SearchService {
     private final SceneRepository repository;
     private final SearchConverter converter;
 
-    public List<Scene> getScenesBy(Map<String, Object> params) throws SQLException {
+    public List<MappedScene> getScenesBy(Map<String, Object> params) throws SQLException {
         return repository.findAllByParamsMap(params);
     }